second pass
